Fix linux natives (#354)
* Remove opus linux-arm64 natives

Unity (unfortunately) does not support(*) linux on arm64.
For some reason, the arm64 native was being chosen above the x64 native
when building the StandaloneLinux64 build of Basis, so let's remove it.

* Fix StandaloneLinux64 choosing x86 version of libopenvr_api.so

BasisVR/Basis#324 broke openvr for the StandaloneLinux64 build
because by adding a meta file for the x86 version of libopenvr_api.so,
it added it for consideration for inclusion with the StandaloneLinux64
build.

Fix this by telling unity to not use it for any platform.